rename file for current naming convention
